+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Fixed
+
+- `SinglePointResults.normal_modes_cartesian` now has the correct length which should be the same as `SinglePointResults.freqs_wavenumber`.
+
 ## [0.12.3] - 2025-01-29
 
 ### Added